Support `owl:Individual` instantiation
