+++ ontogen/__init__.py
@@ -1,2 +1,2 @@
 from .base import Ontology
-from .primitives import OwlClass, OwlProperty, OwlObjectProperty
+from .primitives import OwlClass, OwlObjectProperty
